Move der-focused tests into der.rs since they have nothing to do with signing
I agree to license my contributions to each file under the terms given
at the top of each file I changed.
